pyproject.toml
[project]
name = "skillproof"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.134.0",
    "google-genai>=1.65.0",
    "pillow>=12.1.1",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "qrcode>=8.2",
    "reportlab>=4.4.10",
    "requests>=2.32.5",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.41.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/skillproof"]
